feat: Refactor components
