+++ src/components/About.tsx
@@ -1,0 +1,8 @@
+
+export function About() {
+    return (
+        <div>
+            <h1>Estas en About</h1>
+        </div>
+    );
+}
