+++ src/components/Home.tsx
@@ -1,0 +1,8 @@
+
+export const Home = () => {
+    return (
+        <div>
+            <h1>Hola mundo, estas en Home</h1>            
+        </div>
+    );
+};
