+++ src/components/NavBar.tsx
@@ -1,0 +1,19 @@
+import { NavLink } from 'react-router-dom';
+
+export const NavBar = () => {
+ return (
+ <nav>
+       <ul>
+          <li>
+             <NavLink to="/">Home</NavLink>
+          </li>
+          <li>
+             <NavLink to="/about">About</NavLink>
+          </li>
+          <li>
+             <NavLink to="/products">Products</NavLink>
+          </li>
+       </ul>
+ </nav>
+ );
+};
